Backend/api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi
sqlalchemy
pyodbc
uvicorn[standard]
pydantic
sqlmodel
python-dotenv
azure-cosmos
aiohttp